+++ CHANGELOG.md
@@ -3,15 +3,16 @@
 ## Unreleased
 
 ### Added
-- Added `craft\commerce\elements\Product::isDeletable()`. ([#2606](https://github.com/craftcms/commerce/issues/2606))
-- Added `craft\commerce\controllers\ProductsController::enforceEditProductPermissions()` and `enforceDeleteProductPermissions()`. ([#2606](https://github.com/craftcms/commerce/issues/2606))
+- Added `craft\commerce\controllers\ProductsController::enforceEditProductPermissions()`.
+- Added `craft\commerce\controllers\ProductsController::enforceDeleteProductPermissions()`.
 - Added `craft\commerce\controllers\ProductsPreviewController::enforceEditProductPermissions()`.
-- Added `craft\commerce\elements\Subscription::isEditable()`.
 - Added `craft\commerce\controllers\SubscriptionsController::enforceEditSubscriptionPermissions()`.
 
 ### Changed
 - Improved the performance of order saving.
 - Improved the performance of forumla condition evaluation with result caching.
+- Products now support `EVENT_DEFINE_IS_EDITABLE` and `EVENT_DEFINE_IS_DELETABLE`. ([#2606](https://github.com/craftcms/commerce/issues/2606))
+- Subscriptions now support `EVENT_DEFINE_IS_EDITABLE`.
 
 ## Deprecated
 - Deprecated `craft\commerce\controllers\ProductsController::enforceProductPermissions()`.